Add getAttachments to PDFDocument
This adds the getAttachments method to get all the attachments from a
pdf document.

This is basically implementing the suggestion in
https://github.com/Hopding/pdf-lib/issues/534#issuecomment-662758789

There's some interesting behavior here, as the attachments don't show up
until after save is called. That's tested in the test and described in
the README, just so it's clear that's the behavior.

I don't know how to do all that great jsfiddle setup, so for the README
Ijust copied the basic example from that issue for now.